Rename photon_scattering.py to full_focal_plane.py
