+++ imsim/camera.py
@@ -228,25 +228,26 @@ class Camera(dict):
         if not match:
             # If the given detector is in the list but couldn't be matched, then
             # it's a wavefront or guide detector.
-            raise galsim.GalSimValueError("Arg must be a science detector, not wavefront or guide", det_name)
+            raise galsim.GalSimValueError(
+                "Arg must be a science detector, not wavefront or guide", det_name)
         r00 = (int(match.group(1)[0]), int(match.group(1)[1]))
         s00 = (int(match.group(2)[0]), int(match.group(2)[1]))
 
-        # The S indices aren't too bad. We'll always have a full set of the nine
-        # index pairs, but they need to be permuted to match the position within
-        # the raft.
-        s_rows = np.roll(np.array([0, 1, 2]), -s00[0] + 1)
-        s_cols = np.roll(np.array([0, 1, 2]), -s00[1] + 1)
+        # The following method of manipulating the indices in the detector name
+        # is about 5000 times faster than using lsst.afw.cameraGeom to find the
+        # detector's center and then the names of those shifted away by one
+        # detector width/height.
 
-        # The R indices are more complex. If det_name is positioned to one side
-        # of the raft, we'll need to decrement the lowest index or increment the
-        # highest (and never both) for either or both of the row or column
-        # indices. This integer arithmatic makes it work.
+        # We'll always have a full set of the nine S index pairs, but they need
+        # to be permuted to match the position within the raft.
+        s_rows = np.roll(np.array([0, 1, 2]), 1 - s00[0])
+        s_cols = np.roll(np.array([0, 1, 2]), 1 - s00[1])
+
+        # Use some integer arithmetic to increment or decrement the raft indices
+        # if the detector is positioned on an edge of the raft.
         r_rows = np.array([r00[0]-s_rows[0]//2, r00[0], r00[0]+(2-s_rows[2])//2])
         r_cols = np.array([r00[1]-s_cols[0]//2, r00[1], r00[1]+(2-s_cols[2])//2])
-        
-        # That out of the way, loop through all index combinations and add to the
-        # the list if they exist in the camera.
+
         adjacent_detectors = [f"R{ri}{rj}_S{si}{sj}"
                               for ri, si in zip(r_rows, s_rows)
                               for rj, sj in zip(r_cols, s_cols)
